refactor: Update LimiterAttackHoldRelease struct to use an array of ExponentialRelease objects
The LimiterAttackHoldRelease struct in Limiter.h has been modified to use an array of 8 ExponentialRelease objects instead of a single ExponentialRelease object. This change allows for more flexibility and control over the release behavior. The previous code that initialized a single ExponentialRelease object has been commented out, and a loop has been added to initialize each element of the array with the desired release samples.

This change improves the functionality and performance of the LimiterAttackHoldRelease struct.